requirements.txt
absl-py==1.2.0
asgiref==3.3.1
asttokens==2.1.0
astunparse==1.6.3
backcall==0.2.0
bitarray==2.6.0
blinker==1.6.2
box2d-py==2.3.5
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cipher==0.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
Django==3.1.4
et-xmlfile==1.1.0
executing==1.2.0
Flask==2.3.2
flatbuffers==22.9.24
fonttools==4.36.0
gast==0.4.0
gevent==21.12.0
glob2==0.7
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
greenlet==1.1.3
grpcio==1.49.1
gym==0.26.2
gym-notices==0.0.8
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
imutils==0.5.4
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonpickle==2.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
mahotas==1.4.13
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.4.0
matplotlib-inline==0.1.6
mlxtend==0.21.0
mpi4py==3.1.4
mpmath==1.2.1
Naked==0.1.31
NeedForCryptography==1.5
networkx==2.8.8
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
patsy==0.5.2
pd==0.0.4
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.10.0
prompt-toolkit==3.0.32
protobuf==3.19.6
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.0
pygame-menu==4.3.9
Pygments==2.13.0
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.6
pyparsing==3.0.9
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2020.4
pyvis==0.3.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
scikit-learn==1.1.3
scipy==1.9.0
seaborn==0.12.0
shellescape==3.8.1
six==1.16.0
splunk-sdk==1.6.19
sqlparse==0.4.1
stack-data==0.6.0
statsmodels==0.13.2
swig==4.1.1
sympy==1.11.1
tenacity==8.0.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
threadpoolctl==3.1.0
traitlets==5.5.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.3.4
wrapt==1.14.1
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
